requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py==1.0.0
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
audioread==2.1.9
babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.3.0
bleach==5.0.0
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
curio==1.5
cycler==0.11.0
cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.16.1
imagesize==1.3.0
imgaug==0.4.0
importlib-metadata==4.11.2
importlib-resources==5.6.0
iniconfig==1.1.1
ipykernel==6.12.1
ipyparallel==8.2.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.8.0
keras-preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.9.1
llvmlite==0.38.0
markdown==3.3.6
markupsafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.7.1
notebook==6.4.10
numba==0.55.1
numpy==1.21.0
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
pillow==9.0.1
pip==20.2.4
platformdirs==2.5.1
playsound==1.3.0
pluggy==1.0.0
pooch==1.6.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycparser==2.21
pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-asyncio==0.18.3
python-dateutil==2.8.2
pytz==2022.1
pywavelets==1.3.0
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.3.0
qtpy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
send2trash==1.8.0
setuptools==50.3.0
shapely==1.8.1.post1
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.4.4
soundfile==0.10.3.post1
soupsieve==2.3.2
sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.3.25
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
typing-extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.3
wheel==0.35.1
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.13.3
zipp==3.7.0
